Renaming to app.py to fix command line tile_pages usage
